fix(doctor): gate POSIX-only state-dir tests, fail on an unreadable job store
Address CI (Windows) failure and review feedback.

The two state-dir FAIL tests made Unix-only filesystem assumptions and failed on
Windows: TestDoctorStateDirFreshInstallParentUnwritable relied on mode 0o500
denying writes (Windows ignores POSIX mode bits), and TestDoctorStateDirStatError
relied on os.Stat returning ENOTDIR for a path under a regular file (Windows maps
it to a path-not-found error os.IsNotExist treats as true, taking a different
arm). Both move to doctor_posix_test.go; the cross-platform healthy fresh-install
test stays.

checkStaleJobs now fails, rather than warns, when Store.List returns an error: an
unreadable store root is the fail-closed condition startup RestoreAndCollect and
RestoreGate already refuse to start on, so doctor must exit non-zero rather than
report a healthy exit 0 for an unusable state directory. A missing jobs dir is
not this case (Store.List returns an empty list and no error for a fresh
install). Adds a POSIX test for the unreadable-store path.

Names exit code 2 (usage error) in the doctorMain doc comment.
